+++ Repository/Block.cs
@@ -15,11 +15,11 @@ namespace Repository
         {
         }
 
-        public long index { get; set; }
+        public int index { get; set; }
         public string minedBy { get; set; }
         public string data { get; set; }
         public string previousHash { get; set; }
-        public long nonce { get; set; }
+        public int nonce { get; set; }
         public string hash { get; set; }
     }
 }
